# docs/api/openapi.yaml
openapi: 3.0.3
info:
  title: Pharmacy API
  description: >
    API for pharmacy management system - Users, Items, Orders.
    Items support image upload and retrieval (stored in PostgreSQL as BYTEA).
    Items also support a category field (string).
  version: 1.2.0
  contact:
    name: Pharmacy Team
    email: support@pharmacy.local
  license:
    name: MIT

servers:
  - url: http://localhost:8000
    description: Local development
  - url: https://pharmacy-api-k0ad.onrender.com
    description: Render deployment

tags:
  - name: Health
    description: Service health and basic checks
  - name: Users
    description: Users management
  - name: Items
    description: Pharmacy items management
  - name: Orders
    description: Orders management

paths:
  /hello:
    get:
      tags: [Health]
      summary: Hello World endpoint
      description: Returns a message to verify the backend is running
      responses:
        "200":
          description: Success
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/HelloResponse"

  /health:
    get:
      tags: [Health]
      summary: Health check endpoint
      description: Returns the health status of the backend
      responses:
        "200":
          description: Service is healthy
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/HealthResponse"

  /api/users:
    post:
      tags: [Users]
      summary: Create a new user
      description: Create a new user in the system
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UserCreate"
      responses:
        "201":
          description: User created successfully
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/User"
        "400":
          description: Invalid input
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"

  /api/users/{user_id}:
    get:
      tags: [Users]
      summary: Get user by ID
      description: Retrieve a user by their unique ID
      parameters:
        - $ref: "#/components/parameters/UserId"
      responses:
        "200":
          description: User found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/User"
        "404":
          description: User not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"

  /api/items:
    get:
      tags: [Items]
      summary: List all items
      description: Retrieve a list of all items in the pharmacy
      responses:
        "200":
          description: List of items
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Item"
    post:
      tags: [Items]
      summary: Create a new item
      description: Add a new item (medicine/product) to the pharmacy inventory
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ItemCreate"
      responses:
        "201":
          description: Item created successfully
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Item"
        "400":
          description: Invalid input
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"

  /api/items/{item_id}:
    get:
      tags: [Items]
      summary: Get item by ID
      description: Retrieve a specific item by its ID
      parameters:
        - $ref: "#/components/parameters/ItemId"
      responses:
        "200":
          description: Item found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Item"
        "404":
          description: Item not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
    put:
      tags: [Items]
      summary: Update an item
      description: Update an existing item's details
      parameters:
        - $ref: "#/components/parameters/ItemId"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ItemUpdate"
      responses:
        "200":
          description: Item updated successfully
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Item"
        "404":
          description: Item not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "400":
          description: Invalid input
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
    delete:
      tags: [Items]
      summary: Delete an item
      description: Remove an item from the inventory
      parameters:
        - $ref: "#/components/parameters/ItemId"
      responses:
        "204":
          description: Item deleted successfully
        "404":
          description: Item not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"

  /api/items/{item_id}/image:
    post:
      tags: [Items]
      summary: Upload item image
      description: >
        Upload an image for an item. The image is stored in the database (BYTEA).
        Supported types: image/jpeg, image/png, image/webp.
      parameters:
        - $ref: "#/components/parameters/ItemId"
      requestBody:
        required: true
        content:
          multipart/form-data:
            schema:
              type: object
              required: [file]
              properties:
                file:
                  type: string
                  format: binary
      responses:
        "200":
          description: Image uploaded successfully
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ImageUploadResponse"
        "400":
          description: Invalid file type or empty file
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "404":
          description: Item not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "413":
          description: File too large
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"

    get:
      tags: [Items]
      summary: Get item image
      description: Returns the item's image bytes with the correct content-type.
      parameters:
        - $ref: "#/components/parameters/ItemId"
      responses:
        "200":
          description: Image bytes
          content:
            image/jpeg:
              schema:
                type: string
                format: binary
            image/png:
              schema:
                type: string
                format: binary
            image/webp:
              schema:
                type: string
                format: binary
        "404":
          description: Item not found or image not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"

  /api/orders:
    post:
      tags: [Orders]
      summary: Create a new order
      description: Create an order for a user with multiple items
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/OrderCreate"
      responses:
        "201":
          description: Order created successfully
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Order"
        "400":
          description: Invalid input
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "404":
          description: User or item not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"

  /api/orders/{order_id}:
    get:
      tags: [Orders]
      summary: Get order by ID
      description: Retrieve a specific order by its ID
      parameters:
        - $ref: "#/components/parameters/OrderId"
      responses:
        "200":
          description: Order found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Order"
        "404":
          description: Order not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"

components:
  parameters:
    UserId:
      name: user_id
      in: path
      required: true
      description: User ID (UUID)
      schema:
        type: string
        example: "550e8400-e29b-41d4-a716-446655440000"

    ItemId:
      name: item_id
      in: path
      required: true
      description: Item ID (UUID)
      schema:
        type: string
        example: "550e8400-e29b-41d4-a716-446655440001"

    OrderId:
      name: order_id
      in: path
      required: true
      description: Order ID (UUID)
      schema:
        type: string
        example: "550e8400-e29b-41d4-a716-446655440003"

  schemas:
    HelloResponse:
      type: object
      properties:
        message:
          type: string
          example: "Hello World! Backend is running"

    HealthResponse:
      type: object
      properties:
        status:
          type: string
          example: "ok"

    ErrorResponse:
      type: object
      properties:
        detail:
          type: string
          example: "Item not found"

    User:
      type: object
      required: [user_id, email, name]
      properties:
        user_id:
          type: string
          description: Unique user identifier
          example: "550e8400-e29b-41d4-a716-446655440000"
        email:
          type: string
          format: email
          description: User email address
          example: "john@example.com"
        name:
          type: string
          description: User full name
          example: "John Doe"

    UserCreate:
      type: object
      required: [email, name]
      properties:
        email:
          type: string
          format: email
          description: User email address
          example: "john@example.com"
        name:
          type: string
          description: User full name
          example: "John Doe"

    Item:
      type: object
      required: [item_id, name, price, category]
      properties:
        item_id:
          type: string
          description: Unique item identifier
          example: "550e8400-e29b-41d4-a716-446655440001"
        name:
          type: string
          description: Item name
          example: "Aspirin"
        description:
          type: string
          nullable: true
          description: Item description
          example: "Pain reliever"
        price:
          type: number
          format: double
          description: Item price in USD
          example: 5.99
        category:
          type: string
          description: Item category
          example: "Головная боль"
        has_image:
          type: boolean
          description: True if image exists for the item
          example: true
        image_mime:
          type: string
          nullable: true
          description: Image content type (if image exists)
          example: "image/jpeg"

    ItemCreate:
      type: object
      required: [name, price]
      properties:
        name:
          type: string
          description: Item name
          example: "Aspirin"
        description:
          type: string
          nullable: true
          description: Item description
          example: "Pain reliever"
        price:
          type: number
          format: double
          description: Item price in USD
          example: 5.99
        category:
          type: string
          description: Item category
          example: "Головная боль"
          default: "Головная боль"

    ItemUpdate:
      type: object
      properties:
        name:
          type: string
          nullable: true
          description: Item name
          example: "Aspirin Plus"
        description:
          type: string
          nullable: true
          description: Item description
          example: "Pain reliever (updated)"
        price:
          type: number
          format: double
          nullable: true
          description: Item price in USD
        category:
          type: string
          nullable: true
          description: Item category
          example: "Головная боль"

    ImageUploadResponse:
      type: object
      properties:
        status:
          type: string
          example: "ok"
        item_id:
          type: string
          example: "550e8400-e29b-41d4-a716-446655440001"
        bytes:
          type: integer
          example: 123456
        mime:
          type: string
          example: "image/jpeg"

    Order:
      type: object
      required: [order_id, user_id, created_at, status]
      properties:
        order_id:
          type: string
          description: Unique order identifier
          example: "550e8400-e29b-41d4-a716-446655440003"
        user_id:
          type: string
          description: User who placed the order
          example: "550e8400-e29b-41d4-a716-446655440000"
        created_at:
          type: string
          format: date-time
          description: Order creation timestamp
          example: "2026-01-06T20:00:00Z"
        status:
          type: string
          description: Order status
          example: "pending"

    OrderItemCreate:
      type: object
      required: [item_id, quantity]
      properties:
        item_id:
          type: string
          description: Item ID
          example: "550e8400-e29b-41d4-a716-446655440001"
        quantity:
          type: integer
          minimum: 1
          description: Item quantity
          example: 2

    OrderCreate:
      type: object
      required: [user_id, items]
      properties:
        user_id:
          type: string
          description: User ID
          example: "550e8400-e29b-41d4-a716-446655440000"
        items:
          type: array
          description: List of items in the order
          items:
            $ref: "#/components/schemas/OrderItemCreate"